Set the retrying of slot reservation in a timer instead of every tick
- Add a RetryReservation function in ACC to be called on a timer if the first reservation for a car attempt failed
- Add a condition in Tick to ensure that the car in front of the reservation queue is set to the default speed ceiling since it is in front of the queue
- Set the Phys Material Override to of all ground actors to NonSlippery instead of None
- Add a check to ensure that steering input of a car is below a threshold before reverting speed to the default speed ceiling
- Add BP_ACCORDGameMode to set default pawn to none

+++ Source/ACCORD/Car.h
@@ -35,8 +35,8 @@ public:
 	int GetHeading() const { return Heading; }
 	int GetDirection() const { return Direction; }
 	VehicleProperty* GetVehicleProperty() { return &m_vehicleProperty; }
-	float GetSpeedKPH() const;
-	float GetSpeed() const;
+	double GetSpeedKPH() const;
+	double GetSpeed() const;
 	void SetThrottle(double throttle);
 	void DecreaseThrottle(double throttle);
 	void UpdateHeading();
